build: Core of the library
This is a stripped down build with all options and modules disabled. I will use
this as start configuration for re-engineering the library. Modules and options
can be enabled step by step and tested. First we will leave the source code
unmodified until there are enough Unit Tests that cover it.

Next steps:
 - Make the core build run on MS Windows.
 - Make cmake to alternatively run the core and the default setup.
 - Create Unit Tests for the core code.
 - Enable options and create Unit Tests for them.
 - Modify the source code.