test: Split Unit Tests for old and current code
This way we can test if the old code is still compatible. It will
succeed if no changes were made even with its old bugs. Only a comment
is given for a known bug. Unit Tests on current code will fail on bugs.

In addition export specifications in some header files for unused
program labels are temporary commented and will be reactivated step by
step when they are needed.