test: Create C++ interfaces for system calls
I have tried to overwrite the address of system calls, e.g. fopen(), because I
would not modify the production code. But it didn't work reliably so I decided
to use C++ interfaces as suggested by GoogleTest but in way with minimal
changes to the production code. I only have to include a header file with the
interface class and change the free function system call to a method call of a
class. For example fopen("filename.txt", "a") will become
stdio->fopen("filename.txt", "a").