test: Manage pointer to C++ interfaces
To be able to mock the C++ interfaces there is a global pointer that
points either to the real system functions or to the mock object. Yes,
there has to be taken attention to the global variable but I haven't
found another solution with minimal modification of the production code.
It is needed to Mock old C code but is only a temporary solution as long
as the code isn't completely object oriented.