re-en: Global vars extra header, new ThreadPool if
To have mocking working, compiled in more than one module a
reorganization of the global variables is needed. The so far used
interface classes for the mocking objects are changed to base clases so
only needed mocks can be used and not all virtual methods have to be
overloaded.

Also use the new C++ ThreadPool interface and completely rework and
run gtests using this interface.