test: Create container for fake network interfaces
Use exceptions for error handling on the test tools, with working simple
container test.

Change std::string_view to std::string to be on the save side with
coppying strings because there are problems with raw pointer to the low
level network address structure in the vector container.

It does not work as expected, don't continue.