re-en: Move upnplibInfo and add url class
Move the executable infoprogram from the pupnp subproject to the upnplib
project tree. This is the first step to migrate from the old 'pupnp'
library to the new library 'upnplib'.

Also add a url class to have a url object instead of always handle the
url string. Use the source from https://github.com/homer6/url but modify
it to fit to the library and its policies.

It is also required to adapt the github CI action tests to its upgraded
platforms.