build: Link gtests against upnplib shared library
This is a dead branch. Stopped coding here because of many troubles and
a lot of rework. Keep this only to hold some useful parts and experience
and may be deleted in the future.

In the past all gtests where compiled and linked with just the needed
source files. This meant that source files that were used multiple times
were also recompiled each time. To speed up build time of the gtests
they are now linked against the upnplib shared library.