build: Split build into three general libraries
There are three main builds for the 'upnplib' library:
1. Compatible build with old 'pupnp' library emulation
2. Native 'pupnp' build with old 'pupnp' code
3. Only new 'upnplib' code

All builds are based on new 'upnplib' code but provide downstream
compatibility to the old 'pupnp' API except pos. 3. To manage this I
create three (internal) libraries:
For pos 1.: upnplib_compa
For pos 2.: pupnp_native
For pos 3.: upnplib_native

These libraries are linked according to selected configuration options
to get the resulting library 'upnplib'.

This commit successful build on Linux and MacOS but not on MS Windows so
far.