test: Improve gtests for module 'miniserver'
Also fix a bug in gtest helper class CCaptureStdOutErr to capture output
to stdout and stderr and completely rework it. It is used for the
gtests. The class hasn't closed all needed file descriptors so it
consumes all available fd for the process and silently doesn't output to
stdout/stderr anymore.