test: Use custom matcher for C++ std::regex
This overcomes the mixed internal MatchesRegex() and ContainsRegex()
from Googlemock. On Unix it uses Posix regex but on MS Windows it uses a
limited gmock custom implementation. If using that to be portable we are
limited to the gmock cripple regex for MS Windows.
Reference: https://github.com/google/googletest/issues/1208