test: Use dependency injection for mocking
Rework gtest mocking to use dependency injection and use a static member
variable for the pointer to the injected function. Due to the global
character of the static variable we can also mock functions deeper in
the call stack. Dependency injection effects to all objects from the
injector class because the static variable belongs to the class and not
to the individual object.

This is work in progress with a proof to mock the standard library
function 'select()' with header file <sys/select.h>. The caller class
(injector) 'Sys_select' still contains DEBUG! output messages to have
them available in this commit for possible later use again.