re-en: Move mocking files to upnplib structure
I tried to have two different namespaces for mocking, one '::mocking'
only for the internal pupnp library and one '::upnplib::mocking' only
for the internal upnplib libraries. But I run into serious problems to
address the right namespace. So I decide to use only one namespace
'::upnplib::mocking'. Because it is a new feature for the future it
belongs to the upnplib core project structure.