build: Make mocking run with using two libraries
There is a problem that the injector object (e.g. Stdlib stdlib_h) does
not use the same static pointer variable when testing code on two
libraries 'pupnp_native_shared' and 'upnplib_compa_shared'. The reason
is that the source file containing the object variable (e.g. Stdlib) is
compiled into both libraries to avoid error of missing symbols when
compiling. The error message with missing symbols is a problem of
Microsoft Windows. It requires that global variables for visibility does
not only are decorated with '__declspec(dllexport)'. The consumer must
it also decorate with '__declspec(dllimport) extern'. This issue is
fixed and the source of the injector variable is only compiled one time.