build: Rename 'UPnPlib' project name
The root project name is renamed from 'UPnPlib' to 'UPNPLIB_PROJECT'. We
need the case insensitive name 'upnplib' for the native UPNPLIB source
tree as subproject/subdirectory.

This is part of reorganisation 'pupnp_native' to become 'pupnp' on the
default namespace, 'upnplib_compa' to become 'compa' with namespace
'compa' and 'upnplib_native' to become 'upnplib' with namespace
'upnplib' so that all internal libraries follow the same scheme of
encapsulation.