test: Use Clang 14.0 on Github Action macOS
Also improve gtests for the 'membuffer' module.

On pushing the modified gtests to github I find that Action test
platform 'macOS-11' has been upgraded to 'macOS-12' with default
compiler CLang 14.0. It do not provide compiler 'gcc-10' anymore but
only 'gcc-11'. In the past I have used 'gcc-10' due to feature support
missed with CLang. With the upgrade I decide to use the default CLang
compiler on 'macOS' because it seems that it supports now the latest
updates to the C++ standard. The 'UpNPlib' project is defined to follow
the current standard C++20 for now (may be change to the latest standard
in the future).