build: Reduce pthreads4w include directory entries
Also separate frequently used header files (e.g. visibility.hpp or
port.hpp) from the 'upnplib' include directory so we do not have to
always add it to the seearch path only for this headers. This follows
better the paradigm to only PUBLIC own include directories of a library.
There are header files (e.g. visibility.hpp) that are used in
executables without using the 'upnplib' library.