build: Enable OpenSSL support
This is an interim commit.
It builds on Linux, MacOS and Microsoft Windows but on MacOS one test
fails sometimes with 'bus error' and on WIN32 most gtests fail.

This commit contains the paths and settings to find the libraries of a
system wide OpenSSL installation on Github workflows. It has taken me
some effort to find them so I do not want to lose it.