re-en: Optimize C++ interface for sock module
Define assignment of the interface classes to the namespaces. The
interface class belongs to the 'compa' namespace because it defines a
compatible interface for the old pupnp sock module. The derived class
that calls the old code functions belongs to the namespace 'pupnp'. This
namespace is new created and never used in native pupnp code. Outside
it, it is used to distinguish between the derived interface classes to
the old code (pupnp/sock.hpp) and to the compatible code
(compa/sock.hpp).