re-en: Code first version of SIGPIPE handler
There is a known problem with the SIGPIPE signal that crashes the whole
program if not blocked. But simple blocking with 'signal(SIGPIPE,
SIG_IGN)' is not allowed for a library because it is effective for the
process. It would impact the users program. For more details have a look
at
http://www.microhowto.info/howto/ignore_sigpipe_without_affecting_other_threads_in_a_process.html
and https://stackoverflow.com/q/108183/5014688. Taking the idea and
example from kroki at https://stackoverflow.com/a/2347848/5014688 I code
a handler for this problem. Thanks to kroki.

On Microsoft Windows there is no signal generated so we do not need to
block it as dokumented above.

On MacOS there is the function 'sigtimedwait()' not available that we
need for blocking as noted above. Fortunately we can block it there on
the underlying socket with option SO_NOSIGPIPE.

As you see, we have to handle the issue on each platform separately.