test: Create threaded TLS client and server
This is an interim version from work in progress with commented code
parts, debug messages and time delays to ensure that threads can finish.

Split test for the sock module into tests without threads and with
threads for a simple openssl TLS server because DeathTests together with
threads are unstable. Googletest warn about this.