test: Unstable socket bind on Github Action tests
On the Microsoft Windows platforms on Githup Actions there is a problem
with sockets. With binding to a local interface by selecting ADDR_ANY
there is sometimes a random error message "address already in use" even
if using a single address. I have checked it with a stress test to
provoke the error. I wasn't able to reproduce it with a MS Windows
installation on another hardware so I'm fairly sure that this is only a
github action problem.  The relevant error numbers are WSAEACCES(10013),
WSAEINVAL(10022) and mostly WSAEADDRINUSE(10048):
https://learn.microsoft.com/en-us/windows/win32/winsock/windows-sockets-error-codes-2
Failiing tests due to this issue will be modified step by step.