re-en: CSocket bind with node and port arguments
Change CSocket.bind() argument from addrinfo to node address and port
number as strings. Using strings is to be able to use DNS domain names
and service names in the future. It simplifies managing of the address
family because the address info is now defined by the address family of
the socket. We do not need to check if the address family of the socket
and of the address info always match.