build: compa library direct from source files
Until now the compa library was built with CMake dependency
target_link_library 'pupnp_shared` or `pupnp_static` to use old pupnp
code as long as new code isn't available. This has not proven itself
when linking with old or new code. So we build the compa library also
direct with the old pupnp code if needed, together with the new coded
compa source files.