feat: Platform independent socket error checking
There is a problem that Winsock2 on the Microsoft Windows platform does
not support detailed error information given in the global variable
'errno'. Instead it returns them with calling 'WSAGetLastError()'. I
create a class that encapsulates the difference so there is no need to
always check the platform to get the error information.