re-en: Rework CSigpipe class and make it scoped
On Posix platforms is sigtimedwait() used to disable SIGPIPE for a
thread but that isn't available on macOS/OpenBSD. So I workaround it
with sigpending() and sigwait() to ensure that sigwait() never blocks.

For this I also need:

re-en: Manage global variables in extra files
Create a compile unit 'globalvars.hpp' and globalvars.cpp' for the
compatible code and place the annoing global variables there during
re-engeneering. So we have them all on one place for a better overview
to eleminate them step by step.

test: Rework Unit tests for socket ssl connections