docs: Direct deploy Doxygen html documentation
On its gh-pages Github supports Jekyll by default, a simple blog-aware
application. So uploading content to the gh-pages will build a web site
with JeKyll. This is also true with the simple to use 'docs/' directory
for automatically upload content.

But for Doxygen it is wasted performance because its web pages are
already build to a web site, ready for use. Github provides a workflow
template 'static.yml` that only uploads content without using Jekyll.
This is used now and should increase the upload speed a lot.