docs: Document part of files in Compa/src/genlib/*
Following files are documented:
    Compa/src/genlib/client_table/GenlibClientSubscription.cpp
    Compa/src/genlib/client_table/client_table.cpp
    Compa/src/genlib/miniserver/miniserver.cpp
    Compa/src/genlib/net/http/httpparser.cpp
    Compa/src/genlib/util/strintmap.cpp

Doxyfile has been updated from version 1,9.4 to 1.9.8.

For macOS a sleep of 10 ms is inserted in Utest get captured stderr
output to workaround the empty capture issue.

The draft online documentation has been removed for now due to its size
and amount of files. It will be released when the development
documentation is finished and reduced for productive use.