feat: Setter socket parameter for socket module
Move the initialization arguments for a network socket file descriptor
from the constructor to a separate setter method. This avoids exceptions
within the constructor. It is also possible now to instantiate an object
outside a try scope but set the error prone socket parameter within the
try scope.

Also improve documentation of the socket module.