re-en: Remove setting of Netaddr
Setting a netaddress to a Netaddr object is done with its init() method.
This is removed because we get a verified netaddress only from CAddrinfo
or from a query to local network interfaces.

Improve check for numeric port number.
Adapt Unit tests. Make them not triggering an expensive DNS lookup.